refactor: split public favn wrapper from internal authoring ownership
