+++ apps/favn_authoring/lib/favn/assets.ex
