+++ apps/favn_authoring/lib/favn/public_scaffold.ex
