+++ apps/favn_local/lib/favn/dev/reload.ex
@@ -27,8 +27,8 @@ defmodule Favn.Dev.Reload do
          {:ok, runtime, secrets} <- read_runtime_snapshot(opts),
          :ok <- ensure_no_in_flight_runs(runtime, secrets, opts),
          :ok <- compile_project(),
-         {:ok, build} <- Favn.build_manifest(),
-         {:ok, version} <- Favn.pin_manifest_version(build.manifest),
+         {:ok, build} <- FavnAuthoring.build_manifest(),
+         {:ok, version} <- FavnAuthoring.pin_manifest_version(build.manifest),
          {:ok, runtime_after_restart} <- restart_runner(runtime, secrets, opts),
          :ok <- register_manifest_in_runner(version, runtime_after_restart, secrets),
          :ok <- write_manifest_cache(version, opts),
